pyproject.toml
[tool.poetry]
name = "Hana"
version = "0.1.0"
description = "An automated reddit repost bot selector + purger. Made for r/traaa and other subreddits"
authors = ["No767 <73260931+No767@users.noreply.github.com>"]
license = "GPL-3.0"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "*"
Nuitka = "*"
uvloop = "*"
orjson = "*"
typer = {extras = ["all"], version = "*"}
asyncpraw = "*"
rich = "^12.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
